requirements.txt
inflect==7.5.0
optimum==1.24.0
diffusers==0.32.2
comfyui-frontend-package==1.14.6
torchsde==0.2.6
numpy==1.26.4
einops==0.8.1
transformers==4.50.3
tokenizers==0.21.1
sentencepiece==0.2.0
safetensors==0.5.3
aiohttp==3.11.16
yarl==1.18.3
pyyaml==6.0.2
Pillow==11.1.0
scipy==1.15.2
tqdm==4.67.1
psutil==7.0.0
kornia==0.8.0
spandrel==0.4.1
soundfile==0.13.1
av==14.2.0
importlib_metadata==8.6.1
huggingface_hub==0.30.1
opencv-python==4.11.0.86
filelock==3.16.1
scikit-image==0.25.2
python-dateutil==2.9.0.post0
mediapipe==0.10.21
svglib==1.5.1
fvcore==0.1.5.post20221221
yapf==0.43.0
omegaconf==2.3.0
ftfy==6.3.1
addict==2.4.0
yacs==0.1.8
trimesh[easy]==4.6.6
albumentations==2.0.5
scikit-learn==1.6.1
matplotlib==3.10.1
opencv-python-headless==4.11.0.86
imageio-ffmpeg==0.6.0
lmdb==1.6.2
timm==1.0.15
rich==14.0.0
ultralytics==8.3.40
tyro==0.8.5
dill==0.3.8
color-matcher==0.6.0
mss==10.0.0
facexlib==0.3.0
insightface==0.7.3
onnxruntime==1.21.0
onnxruntime-gpu==1.21.0
accelerate==1.6.0
